# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cloud Edge Backend API",
    "version": "2.4.0",
    "description": "S1-S6 workflow backend and UI-10 management: execution snapshots, guarded logical deletion, ADMIN/USER accounts, image upload/build/distribution, deployments and managed resources. One execution chain; no throughput SDK, long-horizon DQN, physical purge or full Kestra feature parity."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18085"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/namespaces/{namespace}/flows/editor/schema": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "flowEditorSchema",
        "description": "READ. Structural JSON Schema derived from the runtime Flow records. Semantic validation remains server-side; no resource probe or side effects.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Namespace or action forbidden"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/flows/{flowId}/validate": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "flowId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "validateFlowSource",
        "description": "WRITE; active schedule also requires EXECUTE. Parse a single YAML/JSON document using the save validator. No revision, execution, schedule or resource mutation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowDefinition"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Namespace or action forbidden"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/flows/{flowId}/preview": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "flowId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "previewFlowInputs",
        "description": "Same source validation as save; resolve declared input defaults and variables using BindingResolver. Does not execute tasks, expand loops, invent future outputs or guarantee external resource availability.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowPreview"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Namespace or action forbidden"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/flows/{flowId}/export": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "flowId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "exportFlowSource",
        "description": "READ. USER scope only. Source preserves exact saved text; yaml/json serialize the same saved definition without comments. Not a database backup.",
        "parameters": [
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "source",
                "yaml",
                "json"
              ],
              "default": "source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flow source",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/yaml": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowDefinition"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Missing Flow or revision"
          },
          "409": {
            "description": "Different management scope"
          },
          "422": {
            "description": "Invalid format or revision"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/flows/import": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "importFlowSources",
        "description": "WRITE; active schedule requires EXECUTE. Atomic batch of 1..20 unique USER Flows in route namespace, <=1048576 total source characters. Each uses existing expectedRevision CAS; any failure rolls back all. Repeated old CAS returns 409 rather than another revision. Returns flowId-sorted results.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowImportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlowRevision"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Namespace or action forbidden"
          },
          "409": {
            "description": "Revision or management scope conflict; whole batch rolled back"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/flows": {
      "get": {
        "operationId": "listFlows",
        "description": "READ: latest USER flow revision summaries in flowId order. q matches literal id/description substring; labelKey and labelValue must occur together and match exactly.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "labelKey",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "labelValue",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "READ: latest flow revision summaries in flowId order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevisionSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/namespaces/{namespace}/flows/{flowId}": {
      "get": {
        "operationId": "getFlow",
        "description": "READ: source and compiled model, latest unless revision specified",
        "parameters": [
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "READ: source and compiled model, latest unless revision specified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowRevision"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "flowId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "delete": {
        "operationId": "removeFlow",
        "description": "WRITE. Logical USER Flow deletion, CAS expectedRevision; reject active executions/pending afterExecution; remove future schedule atomically. Retain revisions, execution snapshots, idempotency identity. No ID reuse or artifact cleanup.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "parameters": [
          {
            "name": "expectedRevision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/namespaces/{namespace}/flows/{flowId}/revisions": {
      "get": {
        "operationId": "listRevisions",
        "description": "READ: immutable revisions newest first; absent flow returns empty list",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "READ: immutable revisions newest first; absent flow returns empty list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevisionSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "post": {
        "operationId": "saveFlow",
        "description": "WRITE: compare expectedRevision; first save uses 0",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "WRITE: compare expectedRevision; first save uses 0",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowRevision"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "flowId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/namespaces/{namespace}/flows/{flowId}/rollback": {
      "post": {
        "operationId": "rollbackFlow",
        "description": "WRITE: copy target revision into a NEW revision, never overwrite history",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "WRITE: copy target revision into a NEW revision, never overwrite history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowRevision"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "flowId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "FlowSourceRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 262144
          }
        }
      },
      "FlowPreviewRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 262144
          },
          "inputs": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
...
      "FlowDefinition": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "const": 1
          },
          "namespace": {
            "type": "string",
            "pattern": "^[A-Za-z][A-Za-z0-9_.-]{0,99}$"
          },
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z][A-Za-z0-9_.-]{0,99}$"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "labels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "inputs": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Input"
            }
          },
          "variables": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Binding"
            }
          },
          "tasks": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "outputs": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Binding"
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "finally": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "concurrency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Concurrency"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schedule"
              },
              {
                "type": "null"
              }
            ]
          },
          "checks": {
            "type": "array",
            "maxItems": 30,
            "items": {
              "$ref": "#/components/schemas/FlowCheck"
            }
          },
          "webhook": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Opt-in authenticated POST webhook; JSON body contains Flow inputs"
          },
          "sla": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FlowSla"
              },
              {
                "type": "null"
              }
            ]
          },
          "afterExecution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "description": "Durable same-chain tasks after terminal result; failures never rewrite result; inspect TaskRuns"
          }
        },
        "required": [
          "schemaVersion",
          "namespace",
          "id",
          "tasks"
        ],
        "additionalProperties": false
      },
...
    }
  }
}